refactor: move python scripts to folder yolo
