+++ yolo/pyproject.toml
@@ -1,0 +1,49 @@
+[tool.poetry]
+name = "nodejs-yolov5"
+version = "0.1.0"
+description = "A website to detect traffic signs"
+authors = ["Duong Vinh <tienvinh.duong4@gmail.com>"]
+
+[tool.poe.tasks]
+dev = "python yolo.py"
+lint = "flake8"
+"sort_import" = "isort ."
+pretty = "black ."
+
+[tool.isort]
+profile = "black"
+
+[tool.mypy]
+python_version = "3.9"
+ignore_missing_imports = true
+
+[tool.poetry.dependencies]
+python = ">=3.9,<3.12"
+environs = "^9.5.0"
+matplotlib = "^3.5.2"
+numpy = "^1.23.0"
+opencv-python = "^4.6.0.66"
+Pillow = "^9.1.1"
+PyYAML = "^6.0"
+requests = "^2.28.1"
+torch = "^1.12.0"
+torchvision = "^0.13.0"
+tqdm = "^4.64.0"
+tensorboard = "^2.9.1"
+pandas = "^1.4.3"
+seaborn = "^0.11.2"
+scipy = "^1.9.0"
+ipython = "^8.4.0"
+psutil = "^5.9.1"
+thop = "^0.1.1"
+
+[tool.poetry.dev-dependencies]
+poethepoet = "^0.16.0"
+black = {version = "^22.3.0", allow-prereleases = true}
+isort = "^5.10.1"
+flake8 = "^4.0.1"
+mypy = "^0.971"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
